Add files via upload
This project represents the Scarbble game - from the very basic tools and rules, to different ways to use the tools, effective and fast ways to play, and a server - client use of the game. This part of the projects focuses on the Server side. The Client side will be up next.